:sparkles: feat(category-controller/service): userID added as a parameter passed by until repository layer

+++ src/category/controller/categoryController.ts
@@ -19,7 +19,8 @@ class CategoryController {
     req: AuthenticatedRequest,
     res: Response
   ) {
-    const { status, success, message, result } = await CategoryService.getAllCategory();
+    const { userID } = req.user!;
+    const { status, success, message, result } = await CategoryService.getAllCategory(userID);
     res.status(status).json({ success, message, result });
   }
 
@@ -28,7 +29,8 @@ class CategoryController {
     res: Response
   ) {
     const { id } = req.params;
-    const { status, success, message, result } = await CategoryService.getSingleCategory(id);
+    const { userID } = req.user!;
+    const { status, success, message, result } = await CategoryService.getSingleCategory(id, userID);
     res.status(status).json({ success, message, result });
   }
 
@@ -38,8 +40,9 @@ class CategoryController {
   ) {
     const { id } = req.params;
     const { name, color } = req.body;
+    const { userID } = req.user!;
 
-    const { status, success, message, result } = await CategoryService.updatecategory(id, { name, color });
+    const { status, success, message, result } = await CategoryService.updatecategory(id, { name, color }, userID);
     res.status(status).json({ success, message, result });
   }
 
@@ -48,7 +51,8 @@ class CategoryController {
     res: Response
   ) {
     const { id } = req.params;
-    const { status, success, message, result } = await CategoryService.deleteCategory(id);
+    const { userID } = req.user!;
+    const { status, success, message, result } = await CategoryService.deleteCategory(id, userID);
 
     res.status(status).json({ success, message, result });
   }
